% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Sigma.R
\name{Clean_Sigma}
\alias{Clean_Sigma}
\title{Clean_Sigma function}
\usage{
Clean_Sigma(x, cor_method = "s", cor_pairing = "pair")
}
\arguments{
\item{x}{Rhe data used to estimate the correlation matrix. Pairwise correlations between all columns are estimated.}

\item{cor_method}{Method used to estimate correlations. Defaults to Spearman's method. See help file for 'cor' function.}

\item{cor_pairing}{Method used to determine how NA entries are handled. See help file for 'cor' function.}
}
\description{
Primarily a convenience function, Clean_Sigma estimates a correlation (Sigma) matrix that may contain substantial amounts of missing data and returns it in an easily usable form for the glasso function.
}
