% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-igraph.r
\docType{methods}
\name{degree,rnetBasic-method}
\alias{degree,rnetBasic-method}
\title{degree.RnetBasic() - See igraph documentation for full documentation.
#' Gives degrees for all vertices.}
\usage{
\S4method{degree}{rnetBasic}(graph, ...)
}
\arguments{
\item{graph}{an RnetBasic object}

\item{...}{additional arguments passed to igraph::degree()}
}
\description{
degree.RnetBasic() - See igraph documentation for full documentation.
#' Gives degrees for all vertices.
}
