% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-igraph.r
\name{degree}
\alias{degree}
\title{Extended igraph methods}
\usage{
degree(graph, ...)
}
\arguments{
\item{graph}{an igraph class object}

\item{...}{other arguments (see igraph documentation)}
}
\description{
The Rnets package extends the following igraph functions
* modularity
* degree
* gsize
* gorder
* edge_density
}
\details{
degree generic
}
