% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{.Assign_Layout_Matrix}
\alias{.Assign_Layout_Matrix}
\title{Hidden function for assigning layout matrix}
\usage{
.Assign_Layout_Matrix(x)
}
\arguments{
\item{x}{an Rnet object}
}
\description{
Hidden function for assigning layout matrix
}
