% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenR.R
\docType{methods}
\name{.Gen_R}
\alias{.Gen_R}
\alias{.Gen_R,rnetStrata-method}
\title{.Gen_R - internal methods for generating Rnets.}
\usage{
.Gen_R(rnet.obj)

\S4method{.Gen_R}{rnetStrata}(rnet.obj)
}
\arguments{
\item{rnet.obj}{an Rnet S4 object created by Rnet()}
}
\description{
Internal for the "Rnet" methods. .Gen_R should not be called directly.

Internal for the "Rnet" methods. .Gen_R should not be called directly.
}
