% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_e_attrs.R
\docType{data}
\name{E_ATTRS}
\alias{E_ATTRS}
\title{Sample Edge Attributes}
\format{A dataframe from with 4 rows and 2 columns
\describe{
\item{lty}{The line type to use for the 4 edges categories}
\item{width}{The line width to use for the 4 edge categories}

}}
\usage{
data(E_ATTRS)
}
\description{
A small data frame that contains metadata for decorating graph edges using "Assign_Emetadata"
}
