% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_COORDS.R
\docType{data}
\name{EC_COORDS}
\alias{EC_COORDS}
\title{Cartesian vertex coordinates for plotting networks}
\format{A dataframe containing coordinates for vertex locations when plotting networks
\describe{
 \item{vert}{The 3 letter code corresponding to V_ATTRS$Code}
 \item{x}{x-coordinate for positioning vertex in plot}
 \item{y}{x-coordinate for positioning vertex in plot}
}}
\usage{
data(EC_COORDS)
}
\description{
Cartesian vertex coordinates for plotting networks
}
\keyword{datasets}
