% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-igraph.r
\docType{methods}
\name{edge_density,rnetBasic-method}
\alias{edge_density,rnetBasic-method}
\title{edge_density.RnetBasic() - See igraph documentation for full documentation.}
\usage{
\S4method{edge_density}{rnetBasic}(graph, ...)
}
\arguments{
\item{graph}{an RnetBasic object}

\item{...}{additional arguments passed to igraph::gorder()}
}
\description{
edge_density.RnetBasic() - See igraph documentation for full documentation.
}
