% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-igraph.r
\name{edge_density}
\alias{edge_density}
\title{edge_density generic}
\usage{
edge_density(graph, ...)
}
\arguments{
\item{graph}{an igraph class object}

\item{...}{other arguments (see igraph documentation)}
}
\description{
edge_density generic
}
