% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-igraph.r
\docType{methods}
\name{gorder,rnetBasic-method}
\alias{gorder,rnetBasic-method}
\title{gorder.RnetBasic() - See igraph documentation for full documentation.}
\usage{
\S4method{gorder}{rnetBasic}(graph, ...)
}
\arguments{
\item{graph}{an RnetBasic object}

\item{...}{additional arguments passed to igraph::gorder()}
}
\description{
gorder.RnetBasic() - See igraph documentation for full documentation.
}
