% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-igraph.r
\name{gorder}
\alias{gorder}
\title{gorder generic}
\usage{
gorder(graph, ...)
}
\arguments{
\item{graph}{an igraph class object}

\item{...}{other arguments (see igraph documentation)}
}
\description{
gorder generic
}
