% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-igraph.r
\docType{methods}
\name{gsize,rnetBasic-method}
\alias{gsize,rnetBasic-method}
\title{gsize.RnetBasic() - See igraph documentation for full documentation.}
\usage{
\S4method{gsize}{rnetBasic}(graph, ...)
}
\arguments{
\item{graph}{an RnetBasic object}

\item{...}{additional arguments passed to igraph::gsize()}
}
\description{
gsize.RnetBasic() - See igraph documentation for full documentation.
}
