% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-igraph.r
\name{gsize}
\alias{gsize}
\title{gsize generic}
\usage{
gsize(graph, ...)
}
\arguments{
\item{graph}{an igraph class object}

\item{...}{other arguments (see igraph documentation)}
}
\description{
gsize generic
}
