% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{image.edge_heatmap}
\alias{image.edge_heatmap}
\alias{image}
\title{image() method for plotting Rnet heatmaps}
\usage{
\method{image}{edge_heatmap}(x, axes = T, ...)
}
\arguments{
\item{x}{An object with class(x) = 'edge_heatmap', produced by Rnet_Heatmap()}

\item{axes}{logical; if TRUE add axes to the heatmap plot.}

\item{...}{graphical parameters to be passed to image.default() inside this method.}
}
\description{
image() method for plotting Rnet heatmaps
}
