% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-igraph.r
\name{modularity.rnetBasic}
\alias{modularity.rnetBasic}
\alias{modularity}
\title{modularity.RnetBasic() - See igraph documentation for full documentation.}
\usage{
\method{modularity}{rnetBasic}(x, membership, weights = NULL, ...)
}
\arguments{
\item{x}{an}

\item{membership}{Numeric vector, for each vertex it gives its community. The communities are numbered from one.}

\item{weights}{If not NULL then a numeric vector giving edge weights.}

\item{...}{Additional arguments, none currently}
}
\value{
A numeric scalar, the modularity score of the given configuration.
}
\description{
modularity.RnetBasic() - See igraph documentation for full documentation.
}
