% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gomez_modularity_v4_1.R, R/Rnet.R
\docType{methods}
\name{Modularity_Signed}
\alias{Modularity_Signed}
\alias{Modularity_Signed,matrix-method}
\alias{Modularity_Signed,igraph-method}
\alias{Modularity_Signed,rnet.basic-method}
\alias{Modularity_Signed,rnet.strata.multi-method}
\alias{Rnet,ANY,ANY,ANY,ANY,ANY,ANY,ANY,ANY,character-method}
\title{Robust Estimator of modularity (Gomez, et al 2009)}
\usage{
Modularity_Signed(x, membership, weight = NULL)

\S4method{Modularity_Signed}{matrix}(x, membership, weight = NULL)

\S4method{Modularity_Signed}{igraph}(x, membership, weight = NULL)

\S4method{Modularity_Signed}{rnet.basic}(x, membership = NULL,
  weight = "omega")

\S4method{Modularity_Signed}{rnet.strata.multi}(x, membership,
  weight = "omega")

\S4method{Rnet}{ANY,ANY,ANY,ANY,ANY,ANY,ANY,ANY,character}(Data, L1,
  V_set = NULL, n_threshold = 0, cor_method = "s", cor_pairing = "pair",
  Forced_zeros = NULL, Plot_layout = NULL, Stratify = NULL)
}
\arguments{
\item{x}{A graph presented in of the forms discussed below.}

\item{membership}{Defines vertex membership to determine if vertices are similar. May be provided as a string that matches an attribute of x or a vector of length equal to the number of vertices in the graph.}

\item{weight}{Edge weights. Like 'membership', this argument can be defined as a string matching an edge attribute of 'x' or a vector of length equal to the number of edges, but may also be left as NULL which will return an unweighted modularity estimate.}
}
\value{
a numeric value estimating the weighted, signed modularity of x, or a numeric vector containing respective modularity estimates if x contained multiple network.
}
\description{
For flexibility, x may be provided as any of the following formats: an edgelist (data.frame), a weighted adjacency matrix (square numeric matrix), an igraph object, or an rnet.* object (e.g., rnet.basic, rnet.strata.multi, etc.).
}
\details{
Newman's method of estimating graphical modularity based on vertex can accomodate edge weights, but cannot incorporate signed edges, e.g. edges with both positive and negative. Gomez, et al, proposed a similar estimator of modularity estimated in two parts corresponding to positive (Q+) and negative (Q-) edges, and the latter is subtracted from the former. The 'Modularity_Signed' function implements this method of modularity estimation, and returns a scalar.
}
