% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnet_methods.R
\docType{methods}
\name{plot,rnet.basic-method}
\alias{plot,rnet.basic-method}
\title{Plot methods for R-nets}
\usage{
\S4method{plot}{rnet.basic}(x)
}
\arguments{
\item{x}{an rnet object (currently, objects of class 'rnet.basic', 'rnet.strata', 'rnet.multi.strata')}
}
\description{
A plot method for R-nets, and incorporates vertex and edge metadata and layout, if assigns. Only vertex and edge metadata with names that match igraph decoration options (without 'vertex.' or 'edge.' appended to the attribute name; see plot.igraph). Layout is pulled from 'Layout_master' in the rnet object, if it exists. The layout frame can contain 3 columns, with the first column used to match the coordinates in the next two columns to graph vertices OR can contain 2 columns with the same number of vertices in the graph.
}

