% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\docType{methods}
\name{plot,rnet.L1.set,ANY-method}
\alias{plot,rnet.L1.set,ANY-method}
\alias{plot.rnet.L1.set}
\title{plot() method for summarizing L1 StARS section objects}
\usage{
\S4method{plot}{rnet.L1.set,ANY}(x, ylim = c(0, 0.25), ylim_m = c(0,
  0.5), col = "blue", col_m = "red", lty = 1, lty_m = NULL,
  lwd = 2, lwd_m = NULL, mar = c(5, 4, 4, 4), ...)
}
\arguments{
\item{x}{An rnet.L1.set object}

\item{ylim}{plot(ylim) for stability plot}

\item{ylim_m}{plot(ylim) for density plot}

\item{col}{plot(col) for stability plot}

\item{col_m}{plot(col) for density plot#'}

\item{lty}{plot(lty) for stability plot}

\item{lty_m}{plot(lty) for density plot, defaults to lty}

\item{lwd}{plot(lwd) for stability plot}

\item{lwd_m}{plot(lwd) for density plot, defaults to lwd}

\item{mar}{margins for plot.}

\item{...}{Other arguments passed to plot}
}
\description{
plot() method for summarizing L1 StARS section objects
}
