% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\docType{methods}
\name{plot,rnet.strata-method}
\alias{plot,rnet.strata-method}
\title{Plot method for Rnet.strata class}
\usage{
\S4method{plot}{rnet.strata}(x)
}
\arguments{
\item{x}{an rnet object of class 'rnet.strata'}
}
\description{
Plot method for Rnet.strata class
}
