% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\docType{methods}
\name{plot,rnetStrata-method}
\alias{plot,rnetStrata-method}
\title{Plot method for rnetStrata class}
\usage{
\S4method{plot}{rnetStrata}(x)
}
\arguments{
\item{x}{an rnet object of class 'rnetStrata'}
}
\description{
Plot method for rnetStrata class
}
