% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnet_methods.R
\docType{methods}
\name{print,rnet.basic-method}
\alias{print,rnet.basic-method}
\title{Print methods of rnet objects.}
\usage{
\S4method{print}{rnet.basic}(x)
}
\arguments{
\item{x}{An rnet object}
}
\description{
print methods for classes 'rnet.basic', 'rnet.strata', 'rnet.multi.strata'.
}
