% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnet_classes.R
\docType{class}
\name{rnet.L1.set-class}
\alias{rnet.L1.set-class}
\alias{rnet.L1.set}
\title{An S4 class for tracking multiple rnets created from one data set over multiple L1 penalties for comparison and selection.}
\description{
See "L1_selection" documentation for more information.
}
\section{Slots}{

\describe{
\item{\code{Data}}{A dataframe containing the dataset}

\item{\code{L1_set}}{a numeric vector containing the candidate L1 penalties}

\item{\code{B}}{The number of subsamples to draw from the data to evaluate topologic stability}

\item{\code{B_method}}{Assigned either "proportionate" or "Total number" depending on how subsample size is determined.}

\item{\code{B_sets}}{A matrix (n_B x B) containing the rownumbers of the subsamples}

\item{\code{Data_b}}{An arrary (n_b x k x B) containing the data from the B_sets matrix}

\item{\code{pr_b}}{The size of the subsample B as a proportion of the complete dataset}

\item{\code{n_b}}{The size of the subsample}

\item{\code{W_aggr}}{An array (k x k x B x L1) containg all the weighted adjacency matrices generated by the all of subsamples over the L1 penalties}

\item{\code{A_aggr}}{An array (k x k x B x L1) containg all the adjacency matrices generated by the all of subsamples over the L1 penalties}

\item{\code{M}}{A dataframe with with graphical density data over the set of generated networks.}
}}

