% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{rnetSubset-class}
\alias{rnetSubset-class}
\alias{rnetSubset}
\title{An S4 class containing the information of an rnet representing one stratum of data. These objects are used to handle rnet objects and need not be called by the user.}
\description{
Inheirits from 'rnetBasic'
}
\section{Slots}{

\describe{
\item{\code{subset}}{The expression used to define the stratum of data represented by the rnet, e.g., if Statra_def = as.expression(Year == 2008), the rnet is estimated from the subset of data collected in the Year 2008.}
}}

