% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gomez_modularity_v6_1.R
\docType{methods}
\name{signed_modularity,rnetStrata-method}
\alias{signed_modularity,rnetStrata-method}
\title{signed_modularity-rnetMultiStrata}
\usage{
\S4method{signed_modularity}{rnetStrata}(x, membership, weight = "omega")
}
\arguments{
\item{x}{An object of class 'rnetMultiStrata'}

\item{membership}{Defines vertex membership to determine if vertices are similar. May be provided as a string that matches an attribute of x or a vector of length equal to the number of vertices in the graph.}

\item{weight}{Edge weights. Like 'membership', this argument can be defined as a string matching an edge attribute of 'x' or a vector of length equal to the number of edges, but may also be left as NULL which will return an unweighted modularity estimate.}
}
\description{
signed_modularity for class(x) = 'rnetMultiStrata'
}
