% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gomez_modularity_v6_1.R
\docType{methods}
\name{signedModularity,igraph-method}
\alias{signedModularity,igraph-method}
\title{signedModularity for class(x) = 'igraph'}
\usage{
\S4method{signedModularity}{igraph}(x, membership, weight = NULL)
}
\description{
signedModularity for class(x) = 'igraph'
}
