% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gomez_modularity_v6_1.R
\docType{methods}
\name{signedModularity,matrix-method}
\alias{signedModularity,matrix-method}
\title{signedModularity for class(x) = 'matrix'}
\usage{
\S4method{signedModularity}{matrix}(x, membership, weight = NULL)
}
\description{
signedModularity for class(x) = 'matrix'
}
