% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gomez_modularity_v6_1.R
\docType{methods}
\name{signedModularity,rnetBasic-method}
\alias{signedModularity,rnetBasic-method}
\title{signedModularity for class(x) = 'rnetBasic'}
\usage{
\S4method{signedModularity}{rnetBasic}(x, membership = NULL,
  weight = "omega")
}
\description{
signedModularity for class(x) = 'rnetBasic'
}
