% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gomez_modularity_v6_1.R
\docType{methods}
\name{signedModularity,rnetMultiStrata-method}
\alias{signedModularity,rnetMultiStrata-method}
\title{signedModularity for class(x) = 'rnetMultiStrata'}
\usage{
\S4method{signedModularity}{rnetMultiStrata}(x, membership, weight = "omega")
}
\description{
signedModularity for class(x) = 'rnetMultiStrata'
}
