% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-summary.R
\docType{methods}
\name{summary,rnet.L1.set-method}
\alias{summary,rnet.L1.set-method}
\title{Summary - L1 selection object}
\usage{
\S4method{summary}{rnet.L1.set}(object)
}
\arguments{
\item{object}{an object of class 'rnet.L1.set'}
}
\description{
In addition to 'print', this method also shows a table of component D_b values for each edge.
}
