% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-summary.R
\name{summary.mrf_t2}
\alias{summary.mrf_t2}
\title{Summary - MRF comparison with Hotellings T2}
\usage{
summary.mrf_t2(object)
}
\arguments{
\item{object}{an object of class 'mrf_t2'}
}
\description{
Displays crude and adjusted T2 statistics in addition to comparison of the edges in two MRFs
}
