% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnet_methods.R
\docType{methods}
\name{summary,rnet.basic-method}
\alias{summary,rnet.basic-method}
\title{Summary methods of rnet objects.}
\usage{
\S4method{summary}{rnet.basic}(object, ...)
}
\arguments{
\item{x}{An rnet object}
}
\description{
Summary methods for classes 'rnet.basic', 'rnet.strata', 'rnet.multi.strata'. Gives more information than 'print'.
}
