% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\docType{methods}
\name{summary,rnet.basic-method}
\alias{summary,rnet.basic-method}
\alias{summary}
\title{Summary - rnet.basic}
\usage{
\S4method{summary}{rnet.basic}(object)
}
\arguments{
\item{object}{an rnet object of class 'rnet.basic'}

\item{...}{Additional arguments passed to 'summary' method}
}
\description{
Gives more information than 'print'.
}
