% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\docType{methods}
\name{summary,rnet.strata-method}
\alias{summary,rnet.strata-method}
\title{Summary - rnet.strata}
\usage{
\S4method{summary}{rnet.strata}(object)
}
\arguments{
\item{object}{an rnet object of class 'rnet.strata'}
}
\description{
Gives more information than 'print'.
}
