% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-summary.R
\docType{methods}
\name{summary,rnetBasic-method}
\alias{summary,rnetBasic-method}
\alias{summary}
\title{Summary - rnetBasic}
\usage{
\S4method{summary}{rnetBasic}(object)
}
\arguments{
\item{object}{an rnet object of class 'rnetBasic'}

\item{...}{Additional arguments passed to 'summary' method}
}
\description{
Gives more information than 'print'.
}
