% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-summary.R
\docType{methods}
\name{summary,rnetSubset-method}
\alias{summary,rnetSubset-method}
\title{Summary - rnetStrata}
\usage{
\S4method{summary}{rnetSubset}(object)
}
\arguments{
\item{object}{an rnet object of class 'rnetStrata'}
}
\description{
Gives more information than 'print'.
}
