% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnet_methods.R
\docType{methods}
\name{summary,rnet.basic-method}
\alias{summary,rnet.basic-method}
\title{Summary - rnet.basic.}
\usage{
\S4method{summary}{rnet.basic}(object, ...)
}
\arguments{
\item{object}{An rnet object}
}
\description{
Gives more information than 'print'.
}

