% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-igraph.r
\docType{methods}
\name{tkplot,rnetBasic-method}
\alias{tkplot,rnetBasic-method}
\title{tkplot.RnetBasic() - See igraph documentation for full documentation.}
\usage{
\S4method{tkplot}{rnetBasic}(graph, canvas.width = 450,
  canvas.height = 450, ...)
}
\arguments{
\item{graph}{an RnetBasic object}

\item{canvas.width}{tkplot window canvas width}

\item{canvas.height}{tkplot window canvas height}

\item{...}{additional arguments passed to igraph::tkplot()}
}
\description{
tkplot.RnetBasic() - See igraph documentation for full documentation.
}
