% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-igraph.r
\name{tkplot}
\alias{tkplot}
\title{tkplot generic}
\usage{
tkplot(graph, canvas.width = 450, canvas.height = 450, ...)
}
\arguments{
\item{graph}{an igraph class object}

\item{canvas.width}{tkplot window width}

\item{canvas.height}{tkplot window width}

\item{...}{other arguments}
}
\description{
tkplot generic
}
