% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assign_Vmetadata.R
\name{Assign_Vmetadata}
\alias{Assign_Vmetadata}
\title{Assign_Vmetadata - Methods for assigning network vertex metadata}
\usage{
Assign_Vmetadata(network, V_metadata, match.attr = NULL,
  vertex.match.attr = "name", reassign = T)
}
\arguments{
\item{network}{The network to which the vertex metadata will be applied.}

\item{V_metadata}{A dataframe containing the vertex metadata to be assigned. A vertex attribute will be assigned for every column in the frame, except the column used to match V_metadata to existing vertex attribues.}

\item{match.attr}{The name of the column in V_metadata used to match metadata to vertices. Defaults to the first column of V_metadata.}

\item{vertex.match.attr}{the name of the vertex attribute used to match metadata. Defaults to 'name' (V(network)$name), which is typically assigned when the network is created with igraph functions.}

\item{reassign}{A logical argument controling if the function should overwrite the called network argument. Defaults to 'True' for brevity.}
}
\description{
This method assigns metadata to vertices in an igraph object based on an existing vertex attibute, typically vertex name, but can be done with other attributes including those assigned with previous calls with this method.
}
\details{
This method also works with all rnet objects (currently class 'rnet.basic', 'rnet.strata', and 'rnet.multi.strata'), and also adds the names of the metadata attributes to the 'V_metadata' slot.
}
