% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Sigma.R
\name{Clean_Sigma}
\alias{Clean_Sigma}
\title{Clean_Sigma function}
\usage{
Clean_Sigma(x, cor_method = "s", cor_pairing = "pair",
  replace.na = TRUE)
}
\arguments{
\item{x}{Data used to estimate the correlation matrix.}

\item{cor_method}{Character indicating method used to estimate correlations. Defaults to Spearman's method. See help file for 'cor' function.}

\item{cor_pairing}{Character indicating the method used to determine how NA entries are handled. See help file for 'cor' function.}

\item{replace.na}{Logical. Will replace NAs with 0s if TRUE.}
}
\description{
Primarily a convenience function, Clean_Sigma estimates a correlation (Sigma) matrix that may contain substantial amounts of missing data and returns it in an easily usable form for the glasso function.
}
