% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ec_coords.R
\docType{data}
\name{EC_COORDS}
\alias{EC_COORDS}
\title{Example Vertex coordinates}
\format{A dataframe containing coordinates for vertex locations when plotting networks
\describe{
 \item{vert}{The 3 letter code corresponding to V_ATTRS$Code}
 \item{x}{x-coordinate for positioning vertex in plot}
 \item{y}{x-coordinate for positioning vertex in plot}
}}
\usage{
data(EC_COORDS)
}
\description{
A set of x-y coordinates for drawing an R-net for the example E. coli data.
}
