% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-print.R
\docType{methods}
\name{print,rnetBasic-method}
\alias{print,rnetBasic-method}
\alias{print,rnetSubset-method}
\alias{print,rnetStrata-method}
\alias{print,rnet.L1.set-method}
\title{Print methods of rnet objects.}
\usage{
\S4method{print}{rnetBasic}(x)

\S4method{print}{rnetSubset}(x)

\S4method{print}{rnetStrata}(x)

\S4method{print}{rnet.L1.set}(x)
}
\arguments{
\item{x}{An rnet object}
}
\description{
print methods for classes 'rnetBasic', 'rnetStrata', 'rnet.multi.strata'.
}
