% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{rnetStrata-class}
\alias{rnetStrata-class}
\alias{rnetStrata}
\title{An S4 class containing multiple Rnets, each the information of an rnet representing one stratum of data. These objects are used to handle rnet objects and need not be called by the user.}
\description{
Inheirits from 'rnetInput'
}
\section{Slots}{

\describe{
\item{\code{stratify_by}}{The name of the variable used to the original data, e.g. if Stratify_by = 'Year', the object contains 1 rnetSubset object for each unique value of the variable "year" in the dataset.}

\item{\code{E_aggr}}{A matrix containing every edge found in the entire set of rnets, and the stratum in which it was found.}

\item{\code{R_set}}{A list of rnetStrata objects for each strata of the declared Stratify_by variable.}
}}

