% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gomez_modularity_v6_1.R
\docType{methods}
\name{signedModularity,rnetBasic-method}
\alias{signedModularity,rnetBasic-method}
\title{signedModularity-rnetBasic}
\usage{
\S4method{signedModularity}{rnetBasic}(x, membership = NULL,
  weight = "omega")
}
\arguments{
\item{x}{An object of class 'rnetBasic'}

\item{membership}{Defines vertex membership to determine if vertices are similar. May be provided as a string that matches an attribute of x or a vector of length equal to the number of vertices in the graph.}

\item{weight}{Edge weights. Like 'membership', this argument can be defined as a string matching an edge attribute of 'x' or a vector of length equal to the number of edges, but may also be left as NULL which will return an unweighted modularity estimate.}
}
\description{
signedModularity for class(x) = 'rnetBasic'
}
