% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-summary.R
\docType{methods}
\name{summary,rnetMultiStrata-method}
\alias{summary,rnetMultiStrata-method}
\title{Summary - rnetMultiStrata}
\usage{
\S4method{summary}{rnetMultiStrata}(object)
}
\arguments{
\item{object}{an rnet object of class 'rnetMultiStrata'}
}
\description{
Gives more information than 'print'.
}
